pyproject.toml
[project]
name = "focusflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.36.0",
    "matplotlib>=3.10.8",
    "numba>=0.57.0",
    "numpy<2.0.0",
    "onnxruntime",
    "pillow>=12.1.0",
    "protobuf>=6.33.5",
    "pyyaml>=6.0.3",
    "rembg",
    "sentencepiece>=0.2.1",
    "timm>=1.0.20",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
]
